+++ build_runner/lib/src/asset_graph/serialization.dart
@@ -17,7 +17,7 @@ class _AssetGraphDeserializer {
   final Map _serializedGraph;
 
   _AssetGraphDeserializer(List<int> bytes)
-      : _serializedGraph = JSON.decode(UTF8.decode(bytes)) as Map;
+      : _serializedGraph = json.decode(utf8.decode(bytes)) as Map;
 
   /// Perform the deserialization, should only be called once.
   AssetGraph deserialize() {
@@ -74,9 +74,10 @@ class _AssetGraphDeserializer {
 
   AssetNode _deserializeAssetNode(List serializedNode) {
     AssetNode node;
-    var typeId = _NodeType.values[serializedNode[_Field.NodeType.index] as int];
-    var id = _idToAssetId[serializedNode[_Field.Id.index] as int];
-    var serializedDigest = serializedNode[_Field.Digest.index] as String;
+    var typeId =
+        _NodeType.values[serializedNode[_AssetField.NodeType.index] as int];
+    var id = _idToAssetId[serializedNode[_AssetField.Id.index] as int];
+    var serializedDigest = serializedNode[_AssetField.Digest.index] as String;
     var digest = _deserializeDigest(serializedDigest);
     switch (typeId) {
       case _NodeType.Source:
@@ -89,25 +90,30 @@ class _AssetGraphDeserializer {
         break;
       case _NodeType.Generated:
         assert(serializedNode.length == _WrappedGeneratedAssetNode._length);
+        var offset = _AssetField.values.length;
         node = new GeneratedAssetNode(
           id,
-          phaseNumber: serializedNode[_Field.PhaseNumber.index] as int,
-          primaryInput:
-              _idToAssetId[serializedNode[_Field.PrimaryInput.index] as int],
-          state: GeneratedNodeState
-              .values[serializedNode[_Field.State.index] as int],
-          wasOutput:
-              _deserializeBool(serializedNode[_Field.WasOutput.index] as int),
-          builderOptionsId:
-              _idToAssetId[serializedNode[_Field.BuilderOptions.index] as int],
-          globs: (serializedNode[_Field.Globs.index] as Iterable<String>)
+          phaseNumber:
+              serializedNode[_GeneratedField.PhaseNumber.index + offset] as int,
+          primaryInput: _idToAssetId[
+              serializedNode[_GeneratedField.PrimaryInput.index + offset]
+                  as int],
+          state: GeneratedNodeState.values[
+              serializedNode[_GeneratedField.State.index + offset] as int],
+          wasOutput: _deserializeBool(
+              serializedNode[_GeneratedField.WasOutput.index + offset] as int),
+          builderOptionsId: _idToAssetId[
+              serializedNode[_GeneratedField.BuilderOptions.index + offset]
+                  as int],
+          globs: (serializedNode[_GeneratedField.Globs.index + offset]
+                  as Iterable<String>)
               .map((pattern) => new Glob(pattern))
               .toSet(),
           lastKnownDigest: digest,
-          previousInputsDigest: _deserializeDigest(
-              serializedNode[_Field.PreviousInputsDigest.index] as String),
-          isHidden:
-              _deserializeBool(serializedNode[_Field.IsHidden.index] as int),
+          previousInputsDigest: _deserializeDigest(serializedNode[
+              _GeneratedField.PreviousInputsDigest.index + offset] as String),
+          isHidden: _deserializeBool(
+              serializedNode[_GeneratedField.IsHidden.index + offset] as int),
         );
         break;
       case _NodeType.Internal:
@@ -122,11 +128,23 @@ class _AssetGraphDeserializer {
         assert(serializedNode.length == _WrappedAssetNode._length);
         node = new PlaceHolderAssetNode(id);
         break;
+      case _NodeType.PostProcessAnchor:
+        assert(serializedNode.length == _WrappedPostProcessAnchorNode._length);
+        var offset = _AssetField.values.length;
+        node = new PostProcessAnchorNode(
+            id,
+            _idToAssetId[
+                serializedNode[_PostAnchorField.PrimaryInput.index + offset]
+                    as int],
+            serializedNode[_PostAnchorField.ActionNumber.index + offset] as int,
+            _idToAssetId[
+                serializedNode[_PostAnchorField.BuilderOptions.index + offset]
+                    as int]);
     }
     node.outputs.addAll(_deserializeAssetIds(
-        serializedNode[_Field.Outputs.index] as List<int>));
+        serializedNode[_AssetField.Outputs.index] as List<int>));
     node.primaryOutputs.addAll(_deserializeAssetIds(
-        serializedNode[_Field.PrimaryOutputs.index] as List<int>));
+        serializedNode[_AssetField.PrimaryOutputs.index] as List<int>));
     return node;
   }
 
@@ -167,7 +185,7 @@ class _AssetGraphSerializer {
       'assetPaths': assetPaths,
       'failedActions': _serializeFailedActions(_graph.failedActions),
     };
-    return UTF8.encode(JSON.encode(result));
+    return utf8.encode(json.encode(result));
   }
 
   List _serializeNode(AssetNode node) {
@@ -196,21 +214,21 @@ enum _NodeType {
   Generated,
   Internal,
   BuilderOptions,
-  Placeholder
+  Placeholder,
+  PostProcessAnchor,
 }
 
-/// Field indexes for serialized nodes.
-enum _Field {
-  // First, the generic fields for all asset nodes.
+/// Field indexes for all [AssetNode]s
+enum _AssetField {
   NodeType,
   Id,
   Outputs,
   PrimaryOutputs,
-  // **Note**: You must update `_WrappedAssetNode._length` if another generic
-  // field is added below `Digest`.
   Digest,
+}
 
-  // Fields below here are for generated nodes only.
+/// Field indexes for [GeneratedAssetNode]s
+enum _GeneratedField {
   PrimaryInput,
   WasOutput,
   PhaseNumber,
@@ -221,6 +239,13 @@ enum _Field {
   IsHidden,
 }
 
+/// Field indexes for [PostProcessAnchorNode]s.
+enum _PostAnchorField {
+  PrimaryInput,
+  ActionNumber,
+  BuilderOptions,
+}
+
 /// Wraps an [AssetNode] in a class that implements [List] instead of
 /// creating a new list for each one.
 class _WrappedAssetNode extends Object with ListMixin implements List {
@@ -229,7 +254,7 @@ class _WrappedAssetNode extends Object with ListMixin implements List {
 
   _WrappedAssetNode(this.node, this.serializer);
 
-  static final int _length = _Field.Digest.index + 1;
+  static final int _length = _AssetField.values.length;
 
   @override
   int get length => _length;
@@ -240,9 +265,9 @@ class _WrappedAssetNode extends Object with ListMixin implements List {
 
   @override
   Object operator [](int index) {
-    var fieldId = _Field.values[index];
+    var fieldId = _AssetField.values[index];
     switch (fieldId) {
-      case _Field.NodeType:
+      case _AssetField.NodeType:
         if (node is SourceAssetNode) {
           return _NodeType.Source.index;
         } else if (node is GeneratedAssetNode) {
@@ -255,21 +280,23 @@ class _WrappedAssetNode extends Object with ListMixin implements List {
           return _NodeType.BuilderOptions.index;
         } else if (node is PlaceHolderAssetNode) {
           return _NodeType.Placeholder.index;
+        } else if (node is PostProcessAnchorNode) {
+          return _NodeType.PostProcessAnchor.index;
         } else {
           throw new StateError('Unrecognized node type');
         }
         break;
-      case _Field.Id:
+      case _AssetField.Id:
         return serializer._assetIdToId[node.id];
-      case _Field.Outputs:
+      case _AssetField.Outputs:
         return node.outputs
             .map((id) => serializer._assetIdToId[id])
             .toList(growable: false);
-      case _Field.PrimaryOutputs:
+      case _AssetField.PrimaryOutputs:
         return node.primaryOutputs
             .map((id) => serializer._assetIdToId[id])
             .toList(growable: false);
-      case _Field.Digest:
+      case _AssetField.Digest:
         return _serializeDigest(node.lastKnownDigest);
       default:
         throw new RangeError.index(index, this);
@@ -290,9 +317,9 @@ class _WrappedGeneratedAssetNode extends _WrappedAssetNode {
   /// not in [_WrappedAssetNode].
   ///
   /// Indexes below this number are forwarded to `super[index]`.
-  static final int _serializedOffset = _WrappedAssetNode._length;
+  static final int _serializedOffset = _AssetField.values.length;
 
-  static final int _length = _Field.values.length;
+  static final int _length = _serializedOffset + _GeneratedField.values.length;
 
   @override
   int get length => _length;
@@ -304,28 +331,67 @@ class _WrappedGeneratedAssetNode extends _WrappedAssetNode {
   @override
   Object operator [](int index) {
     if (index < _serializedOffset) return super[index];
-    var fieldId = _Field.values[index];
+    var fieldId = _GeneratedField.values[index - _serializedOffset];
     switch (fieldId) {
-      case _Field.PrimaryInput:
+      case _GeneratedField.PrimaryInput:
         return generatedNode.primaryInput != null
             ? serializer._assetIdToId[generatedNode.primaryInput]
             : null;
-      case _Field.WasOutput:
+      case _GeneratedField.WasOutput:
         return _serializeBool(generatedNode.wasOutput);
-      case _Field.PhaseNumber:
+      case _GeneratedField.PhaseNumber:
         return generatedNode.phaseNumber;
-      case _Field.Globs:
+      case _GeneratedField.Globs:
         return generatedNode.globs
             .map((glob) => glob.pattern)
             .toList(growable: false);
-      case _Field.State:
+      case _GeneratedField.State:
         return generatedNode.state.index;
-      case _Field.PreviousInputsDigest:
+      case _GeneratedField.PreviousInputsDigest:
         return _serializeDigest(generatedNode.previousInputsDigest);
-      case _Field.BuilderOptions:
+      case _GeneratedField.BuilderOptions:
         return serializer._assetIdToId[generatedNode.builderOptionsId];
-      case _Field.IsHidden:
+      case _GeneratedField.IsHidden:
         return _serializeBool(generatedNode.isHidden);
+      default:
+        throw new RangeError.index(index, this);
+    }
+  }
+}
+
+/// Wraps a [PostProcessAnchorNode] in a class that implements [List] instead of
+/// creating a new list for each one.
+class _WrappedPostProcessAnchorNode extends _WrappedAssetNode {
+  final PostProcessAnchorNode generatedNode;
+
+  /// Offset in the serialized format for additional fields in this class but
+  /// not in [_WrappedAssetNode].
+  ///
+  /// Indexes below this number are forwarded to `super[index]`.
+  static final int _serializedOffset = _AssetField.values.length;
+
+  static final int _length = _serializedOffset + _PostAnchorField.values.length;
+
+  @override
+  int get length => _length;
+
+  _WrappedPostProcessAnchorNode(
+      this.generatedNode, _AssetGraphSerializer serializer)
+      : super(generatedNode, serializer);
+
+  @override
+  Object operator [](int index) {
+    if (index < _serializedOffset) return super[index];
+    var fieldId = _PostAnchorField.values[index - _serializedOffset];
+    switch (fieldId) {
+      case _PostAnchorField.PrimaryInput:
+        return generatedNode.primaryInput != null
+            ? serializer._assetIdToId[generatedNode.primaryInput]
+            : null;
+      case _PostAnchorField.ActionNumber:
+        return generatedNode.actionNumber;
+      case _PostAnchorField.BuilderOptions:
+        return serializer._assetIdToId[generatedNode.builderOptionsId];
       default:
         throw new RangeError.index(index, this);
     }
@@ -334,9 +400,9 @@ class _WrappedGeneratedAssetNode extends _WrappedAssetNode {
 
 Digest _deserializeDigest(String serializedDigest) => serializedDigest == null
     ? null
-    : new Digest(BASE64.decode(serializedDigest));
+    : new Digest(base64.decode(serializedDigest));
 
 String _serializeDigest(Digest digest) =>
-    digest == null ? null : BASE64.encode(digest.bytes);
+    digest == null ? null : base64.encode(digest.bytes);
 
 int _serializeBool(bool value) => value ? 1 : 0;