+++ build_runner/lib/src/generate/phase.dart
@@ -37,13 +37,31 @@ abstract class BuildPhase {
   int get identity;
 }
 
-/// A [BuildPhase] that uses a single [Builder] to generate files.
-class InBuildPhase extends BuildPhase {
-  final Builder builder;
+/// An "action" in the build graph which represents running a single builder
+/// on a set of sources.
+abstract class BuildAction {
+  /// Either a [Builder] or a [PostProcessBuilder].
+  dynamic get builder;
+  String get builderLabel;
+  BuilderOptions get builderOptions;
+  InputMatcher get generateFor;
+  String get package;
+  InputMatcher get targetSources;
+}
 
+/// A [BuildPhase] that uses a single [Builder] to generate files.
+class InBuildPhase extends BuildPhase implements BuildAction {
+  @override
+  final Builder builder;
+  @override
+  final String builderLabel;
+  @override
   final BuilderOptions builderOptions;
+  @override
   final InputMatcher generateFor;
+  @override
   final String package;
+  @override
   final InputMatcher targetSources;
 
   @override
@@ -54,6 +72,7 @@ class InBuildPhase extends BuildPhase {
   InBuildPhase._(this.package, this.builder, this.builderOptions,
       {@required this.targetSources,
       @required this.generateFor,
+      @required this.builderLabel,
       bool isOptional,
       bool hideOutput})
       : this.isOptional = isOptional ?? false,
@@ -73,6 +92,7 @@ class InBuildPhase extends BuildPhase {
   factory InBuildPhase(
     Builder builder,
     String package, {
+    String builderKey,
     InputSet targetSources,
     InputSet generateFor,
     BuilderOptions builderOptions,
@@ -86,6 +106,9 @@ class InBuildPhase extends BuildPhase {
     return new InBuildPhase._(package, builder, builderOptions,
         targetSources: targetSourceMatcher,
         generateFor: generateForMatcher,
+        builderLabel: builderKey == null || builderKey.isEmpty
+            ? _builderLabel(builder)
+            : _simpleBuilderKey(builderKey),
         isOptional: isOptional,
         hideOutput: hideOutput);
   }
@@ -141,18 +164,29 @@ class PostBuildPhase extends BuildPhase {
 
 /// Part of a larger [PostBuildPhase], applies a single
 /// [PostProcessBuilder] to a single [package] with some additional options.
-class PostBuildAction {
+class PostBuildAction implements BuildAction {
+  @override
   final PostProcessBuilder builder;
+  @override
+  final String builderLabel;
+  @override
   final BuilderOptions builderOptions;
+  @override
   final InputMatcher generateFor;
+  @override
   final String package;
+  @override
   final InputMatcher targetSources;
 
   PostBuildAction(this.builder, this.package,
-      {@required BuilderOptions builderOptions,
+      {String builderKey,
+      @required BuilderOptions builderOptions,
       @required InputSet targetSources,
       @required InputSet generateFor})
-      : builderOptions = builderOptions ?? const BuilderOptions(const {}),
+      : builderLabel = builderKey == null || builderKey.isEmpty
+            ? _builderLabel(builder)
+            : _simpleBuilderKey(builderKey),
+        builderOptions = builderOptions ?? const BuilderOptions(const {}),
         targetSources = new InputMatcher(targetSources ?? const InputSet()),
         generateFor = new InputMatcher(generateFor ?? const InputSet());
 
@@ -162,6 +196,23 @@ class PostBuildAction {
         package,
         targetSources,
       ]);
+}
+
+/// If we have no key find a human friendly name for the Builder.
+String _builderLabel(Object builder) {
+  var label = '$builder';
+  if (label.startsWith('Instance of \'')) {
+    label = label.substring(13, label.length - 1);
+  }
+  return label;
+}
+
+/// Change "angular|angular" to "angular".
+String _simpleBuilderKey(String builderKey) {
+  if (!builderKey.contains('|')) return builderKey;
+  var parts = builderKey.split('|');
+  if (parts[0] == parts[1]) return parts[0];
+  return builderKey;
 }
 
 final _deepEquals = const DeepCollectionEquality();