+++ e2e_example/test/common/utils.dart
@@ -101,12 +101,12 @@ Future<Null> _startServer(String command, List<String> buildArgs,
 
   _process = await Process.start(command, buildArgs);
   _stdOutLines = _process.stdout
-      .transform(UTF8.decoder)
+      .transform(utf8.decoder)
       .transform(const LineSplitter())
       .asBroadcastStream();
 
   _stdErrLines = _process.stderr
-      .transform(UTF8.decoder)
+      .transform(utf8.decoder)
       .transform(const LineSplitter())
       .asBroadcastStream();
 
@@ -172,12 +172,11 @@ Future<Null> _resetGitClient() async {
 }
 
 Future<Null> get nextSuccessfulBuild async {
-  await _stdOutLines
-      .firstWhere((line) => line.contains('Build: Succeeded after'));
+  await _stdOutLines.firstWhere((line) => line.contains('Succeeded after'));
 }
 
 Future<Null> get nextFailedBuild async {
-  await _stdErrLines.firstWhere((line) => line.contains('Build: Failed after'));
+  await _stdErrLines.firstWhere((line) => line.contains('Failed after'));
 }
 
 Future<String> nextStdErrLine(String message) =>