+++ e2e_example/test/goldens/generated_build_script.dart
@@ -14,8 +14,10 @@ final _builders = [
       _i1.toDependentsOf('provides_builder'),
       hideOutput: true,
       appliesBuilders: ['provides_builder|some_post_process_builder']),
-  _i1.apply('build_test|test_bootstrap',
-      [_i3.debugTestBuilder, _i3.testBootstrapBuilder], _i1.toRoot(),
+  _i1.apply(
+      'build_test|test_bootstrap',
+      [_i3.debugIndexBuilder, _i3.debugTestBuilder, _i3.testBootstrapBuilder],
+      _i1.toRoot(),
       hideOutput: true,
       defaultGenerateFor: const _i4.InputSet(include: const ['test/**'])),
   _i1.apply(